hakaton/requirements.txt
streamlit>=1.36,<2.0
requests>=2.32,<3.0
pandas>=2.2.3,<3.0
numpy>=1.26.4
python-dotenv>=1.0,<2.0
Pillow>=10.4,<13.0
matplotlib>=3.9,<4.0
opencv-python-headless>=4.9,<5.0
